+++ library/CHANGELOG.md
@@ -10,7 +10,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - Add ODS typography for iOS & Android platforms ([#121](https://github.com/Orange-OpenSource/ods-flutter/issues/121))
-- Add OdsDivider widget ([#121](https://github.com/Orange-OpenSource/ods-flutter/issues/121))
+- Add `OdsDivider` widget ([#121](https://github.com/Orange-OpenSource/ods-flutter/issues/121))
+- Add `OdsVerticalImageFirstCard` widget ([#8](https://github.com/Orange-OpenSource/ods-flutter/issues/8))
 
 ### Removed